Fixed some missing renamed files.
