+++ src/modules/IEC61131-3/Arithmetic/CMakeLists.txt
@@ -21,10 +21,10 @@ forte_add_sourcefile_hcpp(F_DIV_fbt)
 forte_add_sourcefile_hcpp(F_MOD_fbt)
 forte_add_sourcefile_hcpp(F_EXPT_fbt)
 forte_add_sourcefile_hcpp(F_MOVE_fbt)
-forte_add_sourcefile_hcpp(GEN_F_MOVE_fct)
+forte_add_sourcefile_hcpp(GEN_F_MOVE_fct_fbt)
 forte_add_sourcefile_hcpp(F_MULTIME_fbt)
 forte_add_sourcefile_hcpp(F_DIVTIME_fbt)
-forte_add_sourcefile_hcpp(GEN_ADD)
+forte_add_sourcefile_hcpp(GEN_ADD_fbt)
 forte_add_sourcefile_hcpp(F_TRUNC_fbt)
 forte_add_sourcefile_hcpp(F_ADD_DT_TIME_fbt)
 forte_add_sourcefile_hcpp(F_ADD_TOD_TIME_fbt)
