+++ src/modules/IEC61131-3/BitwiseOperators/genbitbase_fbt.h
